Implement finally job method and move TesseractJob to a different file
